finished

+++ 2023/programacao_orientada_a_objetos/ap2/src/main/java/ap2/com/ap2/services/ClientService.java
@@ -1,7 +1,10 @@
 package ap2.com.ap2.services;
 
+import ap2.com.ap2.Exceptions.NotFoundClient;
+import ap2.com.ap2.dtos.ClientResponseDTO;
 import ap2.com.ap2.entities.Client;
 import ap2.com.ap2.repository.ClientRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.List;
@@ -11,19 +14,33 @@ public class ClientService {
     private ClientRepository clientRepository;
     @Autowired
     public ClientService(ClientRepository clientRepository){this.clientRepository = clientRepository;}
-    public List<Client> getAll(Integer age){
-        return this.clientRepository.findAll();
+    public List<ClientResponseDTO> getAll(Integer age){
+        List<Client> listClients = this.clientRepository.findAll();
+        if(listClients.isEmpty()) throw new NotFoundClient();
+        ClientResponseDTO listClientsDto =  new ClientResponseDTO();
+        return listClientsDto.convertListToClientInDto(listClients);
     }
-    public Optional<Client> getById(int clientId){
-        return this.clientRepository.findById(clientId);
+    public ClientResponseDTO getById(int clientId){
+        Optional<Client> client = this.clientRepository.findById(clientId);
+        if(client.isEmpty()) throw new NotFoundClient();
+        ClientResponseDTO clientDTO = new ClientResponseDTO();
+        return clientDTO.convertToDto(client);
     }
-    public Client store(Client clientData){
-      return this.clientRepository.save(clientData);
+    public String store(Client clientData){
+      this.clientRepository.save(clientData);
+      return "Cliente adicionado com sucesso!";
     }
-    public Client update(Client clientData){
-        return this.clientRepository.save(clientData);
+    @Transactional
+    public String update(Client clientData, int clientId){
+        Optional<Client> client = this.clientRepository.findById(clientId);
+        if(client.isEmpty()) throw new NotFoundClient();
+        this.clientRepository.update(clientData.getName(), clientData.getAge(), clientData.getJob(),clientId);
+        return "Cliente atualizado com sucesso!";
     }
-    public void delete(int clientId){
+    public String delete(int clientId){
+        Optional<Client> client = this.clientRepository.findById(clientId);
+        if(client.isEmpty()) throw new NotFoundClient();
         this.clientRepository.deleteById(clientId);
+        return "Cliente deletado com sucesso!";
     }
 }
